Rename i.b.{BitsquareArgumentParser=>ArgumentParser}
